backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pillow==10.1.0
python-magic==0.4.27
redis==5.0.1
boto3==1.34.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
qrcode[pil]==7.4.2
python-docx==1.1.0
openpyxl==3.1.2
xlsxwriter==3.1.9
ffmpeg-python==0.2.0
pyzbar==0.1.9

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2